Move Pokemon repositories to shared folder in order to avoid code duplication
