+++ src/main/java/spyra/lukasz/pokerestapi/shared/repository/PokeAbilityRepository.java
@@ -1,11 +1,11 @@
-package spyra.lukasz.pokerestapi.consume;
+package spyra.lukasz.pokerestapi.shared.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import spyra.lukasz.pokerestapi.shared.PokeAbility;
 
 import java.util.Optional;
 
-interface PokeAbilityRepository extends JpaRepository<PokeAbility, Long> {
+public interface PokeAbilityRepository extends JpaRepository<PokeAbility, Long> {
 
     Optional<PokeAbility> findFirstByNameAndHidden(String name, boolean hidden);
