+++ src/main/java/spyra/lukasz/pokerestapi/shared/repository/PokeStatRepository.java
@@ -1,11 +1,11 @@
-package spyra.lukasz.pokerestapi.consume;
+package spyra.lukasz.pokerestapi.shared.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import spyra.lukasz.pokerestapi.shared.PokeStat;
 
 import java.util.Optional;
 
-interface PokeStatRepository extends JpaRepository<PokeStat, Long> {
+public interface PokeStatRepository extends JpaRepository<PokeStat, Long> {
 
     Optional<PokeStat> findFirstByNameAndValue(String name, int value);
