+++ src/main/java/spyra/lukasz/pokerestapi/shared/repository/PokeTypeRepository.java
@@ -1,11 +1,11 @@
-package spyra.lukasz.pokerestapi.consume;
+package spyra.lukasz.pokerestapi.shared.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import spyra.lukasz.pokerestapi.shared.PokeType;
 
 import java.util.Optional;
 
-interface PokeTypeRepository extends JpaRepository<PokeType, Long> {
+public interface PokeTypeRepository extends JpaRepository<PokeType, Long> {
 
     Optional<PokeType> findFirstByName(String name);
